requirements.txt
Flask>=2.3,<3.0
Flask-SQLAlchemy>=3.1,<4.0
PyMySQL>=1.1,<2.0
Flask-Login>=0.6,<1.0
openpyxl>=3.1,<4.0
